refactor: Renamed `route_utils` to `gh_route_utils`
