+++ public/index.tsx
@@ -6,7 +6,7 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import AppWrapper from "../website/components/AppWrapper";
 import LandingPage from "../website/components/LandingPage";
 import ErrorPage from "../website/components/ErrorPage";
-import { isQueryStringPath, convertQueryStringPathToUrl } from "../website/utils/route_utils";
+import { isQueryStringPath, convertQueryStringPathToUrl } from "../website/utils/gh_route_utils";
 import StyleProvider from "../src/aics-image-viewer/components/StyleProvider";
 import "../src/aics-image-viewer/assets/styles/typography.css";
 import "./App.css";
